add multi-thread sync

+++ globlItem.h
@@ -8,10 +8,10 @@ typedef unsigned char BYTE;
 //包头
 typedef struct _header
 {
-	UINT		len;	//数据长度
+	size_t		len;	//数据长度
 	BYTE		type;	//包类型
 	BYTE		subtype;
-	UINT        packNum; //包序号
+	size_t      packNum; //包序号
 }HDR, *PHDER;
 
 const int HEADER_SIZE = sizeof(HDR);
